refactor: 채팅방 입장 훅을 채팅방 엔티티로 이동 (#99)
